# Speed dict lookup by caching computed hash value in dictEntry