--- a/src/dict.c
+++ b/src/dict.c
@@ -205,8 +205,9 @@ int dictRehash(dict *d, int n) {
             uint64_t h;
 
             nextde = de->next;
+            de->hash = dictHashKey(d, de->key);            
             /* Get the index in the new hash table */
-            h = dictHashKey(d, de->key) & d->ht[1].sizemask;
+            h = de->hash & d->ht[1].sizemask;
             de->next = d->ht[1].table[h];
             d->ht[1].table[h] = de;
             d->ht[0].used--;
@@ -299,7 +300,8 @@ dictEntry *dictAddRaw(dict *d, void *key, dictEntry **existing)
 
     /* Get the index of the new element, or -1 if
      * the element already exists. */
-    if ((index = _dictKeyIndex(d, key, dictHashKey(d,key), existing)) == -1)
+    uint64_t hash = dictHashKey(d,key);
+    if ((index = _dictKeyIndex(d, key, hash, existing)) == -1)
         return NULL;
 
     /* Allocate the memory and store the new entry.
@@ -314,6 +316,7 @@ dictEntry *dictAddRaw(dict *d, void *key, dictEntry **existing)
 
     /* Set the hash entry fields. */
     dictSetKey(d, entry, key);
+    entry->hash = hash;
     return entry;
 }
 
@@ -376,7 +379,7 @@ static dictEntry *dictGenericDelete(dict *d, const void *key, int nofree) {
         he = d->ht[table].table[idx];
         prevHe = NULL;
         while(he) {
-            if (key==he->key || dictCompareKeys(d, key, he->key)) {
+            if (key==he->key || (he->hash==h && dictCompareKeys(d, key, he->key)) ) {
                 /* Unlink the element from the list */
                 if (prevHe)
                     prevHe->next = he->next;
@@ -485,7 +488,7 @@ dictEntry *dictFind(dict *d, const void *key)
         idx = h & d->ht[table].sizemask;
         he = d->ht[table].table[idx];
         while(he) {
-            if (key==he->key || dictCompareKeys(d, key, he->key))
+            if (key==he->key || (he->hash==h && dictCompareKeys(d, key, he->key)) )
                 return he;
             he = he->next;
         }
@@ -1005,7 +1008,7 @@ static long _dictKeyIndex(dict *d, const void *key, uint64_t hash, dictEntry **e
         /* Search if this slot does not already contain the given key */
         he = d->ht[table].table[idx];
         while(he) {
-            if (key==he->key || dictCompareKeys(d, key, he->key)) {
+            if (key==he->key || (he->hash==hash && dictCompareKeys(d, key, he->key)) ) {
                 if (existing) *existing = he;
                 return -1;
             }