@oranagra 2020-06-12
@tryfinally nice idea, but considering the extra memory usage, i'm not certain it would really worth it.
you're essentially reducing the need for string compare on many cases, however many times the string compare will exit rather quickly (possibly after the first character).
the main improvement might be that your approach is more cache friendly (the hash cached inside the dict entry would in the CPU cache, while the key will not).
however, increasing the size of the dict entry (other than consuming a lot of memory in some use cases), would also reduce the cache efficiency for some use cases.
i'm guessing a benchmark of redis rather than just the dict can show different results (not sure it would be measurable, in which case the memory usage will be a negative impact).
also, even doing a benchmark of just dict (not networking etc), needs to be done on large database that consumes a lot of memory, and test it on big and small dictionaries, filled with big or small keynames/values. i.e. each of these scenarios can have a different result.