@tryfinally 2020-06-13
@oranagra you got it right, the motivation is cache friendliness. 
Most CPUs caches are 64, the jemalloc 32 bytes bin will be used, the memory access will fast especially that no dictEntry will straddle 2 cache line, which can happen with 24bytes dictEntry.
Regarding memory consumptions, the dict-benchmark memory foot print with 5M keys is increased by 40MB which is an increase of 8.84% in mapped memory. 
IMHO the benefit of speeding up dict outweighs the 8.8% increase in memory for large dicts.