@oranagra 2020-06-13
Come to think of it, it'll only reduce the need for string compare for hash collisions. 
Since we always strive to have the hash table bigger than the number of entries, we usually don't have many collisions. 
Anyway in order to really consider this we need to look at more benchmarks, ones that use real redis workload, I.e. Not just doing the lookup, but also using the data (which mixes up the cache), preferably also see that on real redis sever (network), and using big and small key names, and a lot of them (so that we rally use (access) the majority of the machine's ram, and certainly much more than the cache). make  sure there's no rehashing in the test, and make sure to mention the memory increase, not just the benefits.