@oranagra 2020-06-13
Nothing specific. I would probably use memtier_benchmark (with pipeline to somewhat eliminate the network effect). But the main effort is to plan the list of tests, and make sure you don't have rehashing in your test (you can use DEBUG HTSTATS). 
I'm not certain it would be measurable, so it would be interesting to also see a dict benchmark that does the same things (including actual read of the values), But that would be in some way a wasted effort, since I doubt we'll wanna merge it if it wont show a considerable effect in real benchmark, considering the memory usage implications.