@tryfinally 2020-06-13
@oranagra BTW, why not replace dict current implementation that uses chaining with open addressing? that will give more improvements (cache wise).
Is "chaining" used for it's simplicity, especially rehashing implementation?