@oranagra 2020-06-13
Yes, gradual rehashing and the support for SCAN (stateless iterator) would be a problem. 
There's a long term plan to replace the dict with a radix tree (rax.c), which is more memory efficient and also cache friendly. One of the complications is the SCAN command.