@oranagra 2020-06-13
@tryfinally What do you mean? Do you have a way to support SCAN in radix tree?? I'm only explaining why the other change you suggested is problematic.